rename mip to pipeline
